feat: client certificate is accepted
From now on it is possible to inject a client certificate, as well as its key to install it as part of the accepted certificates. It has also been added the possibility to skip the verification in case of using self-signed certificates or certificates on which the name or IP cannot be verified.
Improved testing by checking the cipher SSL associated with database
session.

[#180913283](https://www.pivotaltracker.com/story/show/180913283)