fix: Go 1.22 broke compatibility with Mysql 5.7 (#76)
* fix: Go 1.22 broke compatibility with Mysql 5.7

[#187028341](https://www.pivotaltracker.com/story/show/187028341)

Apparently, mysql tag `5.7` is equivalent to `5.7-oracle`.
Also, it seems that there is an alternative tag named `5.7-debian`.
I ran some tests locally and they return different results.
As such, fixing this issue (which might occur when running
against one image but not against another) is not very useful.

Still, this change tries to workaround the issue for now by changing
the Docker image we use. However, we should tests against a real
AWS instance to really check if the TLS issue also happens in AWS.

* fix: check against allowed encryption methods indistinctly

[#187028341](https://www.pivotaltracker.com/story/show/187028341)

Different OS, DB versions, etc. have different preferences in terms of TLS encryption method.
It is possible that checking whether the encryption method exactly matches one of many similar ones
is not needed.

I leave that discussion open to anyone reviewing this PR.

* chore: apply suggestions from code review